Inclusão de tabelas dinâmicas e também de novos campos dentro do formulário de envio de mensagens.

+++ src/main/java/com/sistema/repository/CadastroPessoasRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.sistema.models.CadastroPessoas;
 
 public interface CadastroPessoasRepository extends CrudRepository<CadastroPessoas, String> {
-	CadastroPessoas findByCodigo (long codigo);
+	CadastroPessoas findByCodigo (int codigo);
 }
